requirements.txt
# PDF Search System Requirements
# Core dependencies for PDF processing and search functionality

# PDF processing
PyPDF2==3.0.1
pymupdf==1.23.26

# Machine learning and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2

# Data processing
numpy==1.24.4
pandas==2.1.4

# Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Natural language processing
nltk==3.8.1

# Additional utilities
typing-extensions==4.8.0
requests==2.31.0

# Production server (optional)
gunicorn==21.2.0
